requirements.txt
# Core dependencies
jieba>=0.42.1
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Topic Modeling
gensim>=4.3.0
scikit-learn>=1.3.0

# Deep Learning & NLP (CPU version)
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
sentencepiece>=0.1.99

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pyLDAvis>=3.4.0
wordcloud>=1.9.0

# Utilities
tqdm>=4.65.0
tqdm>=4.65.0
pyyaml>=6.0
python-dateutil>=2.8.0
regex>=2023.0

# Report Generation
markdown>=3.5.0
jinja2>=3.1.0
weasyprint>=60.0
pypandoc>=1.11

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.4.0
